Rewrote ext2/fat fread functions to use fs_fread Initial commit of nofs, based on an idea by https://github.com/JamesC1 Fixed a bug in emmc.c so that multi block reads now work
git-svn-id: file:///var/johncronin-svn/rpi-boot@97 0f3ef04d-8e58-4d2a-b558-b31704943fbb

+++ block.h
@@ -44,8 +44,8 @@ struct block_device {
 	struct fs *fs;
 };
 
-int block_read(struct block_device *dev, uint8_t *buf, size_t buf_size, uint32_t starting_block);
-int block_write(struct block_device *dev, uint8_t *buf, size_t buf_size, uint32_t starting_block);
+size_t block_read(struct block_device *dev, uint8_t *buf, size_t buf_size, uint32_t starting_block);
+size_t block_write(struct block_device *dev, uint8_t *buf, size_t buf_size, uint32_t starting_block);
 
 #endif
